I fell right into the hole I digged Java is not your friend :-) 1.11131720435081E14 vs 1.1113172043508E14

+++ src/main/java/be/ccfun/day18/CubicMeter.java
@@ -18,11 +18,11 @@ public class CubicMeter {
 	}
 
 	public int getRow() {
-		return position.row();
+		return (int) position.row();
 	}
 
 	public int getCol() {
-		return position.col();
+		return (int) position.col();
 	}
 
 	public String getColor() {
