+++ src/main/java/be/ccfun/day18/Position.java
@@ -1,5 +1,5 @@
 package be.ccfun.day18;
 
-public record Position(int row, int col) {
+public record Position(long row, long col) {
 
 }
